requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu113

torch==1.11.0+cu113
torchvision==0.12.0+cu113
opencv-python==4.6.0.66
torchsummary==1.5.1
tqdm==4.64.0
tensorboard==2.9.1
numpy==1.22.3
Pillow==9.1.1
albumentations==1.1.0
imagesize==1.3.0
scipy==1.8.0